fix: lint issue on dummy-wallet-app
